requirements.txt
Pillow==9.4.0
Pillow==9.5.0
streamlit==1.13.0
streamlit==1.22.0
